refactor(session): extract class redefinition into src/session/jvm/
`redefineClasses` and `RedefineClassesResult` move to
`RedefineClassesController`, which takes the breakpoint controller as a
constructor argument for the post-swap re-send. Its `anchorResolution` field
now names the `AnchorResolution` type instead of repeating its shape; the
result type is re-exported from `session-manager-operations.ts`.
